Apply code review fixes: extract resolve_commit_vaddr helper, add DEFAULT_POOL_SIZE constant, improve comments, fix type consistency
Agent-Logs-Url: https://github.com/Pryancito/eclipse/sessions/9bc42374-8656-4be8-8507-5e1cbc86d4b5

+++ eclipse-apps/lunas/src/state.rs
@@ -298,8 +298,8 @@ impl LunasState {
                                 if w_idx < self.space.window_count && vaddr != 0 {
                                     let win = &mut self.space.windows[w_idx];
                                     win.buffer_handle = Some(vaddr as u64);
-                                    win.w = width;
-                                    win.h = height + ShellWindow::TITLE_H;
+                                    win.w = width as i32;
+                                    win.h = height as i32 + ShellWindow::TITLE_H;
                                 }
                             }
                         }
